mercado_btc_monitor/requirements.txt
fastapi==0.61.1
gunicorn==20.0.4
loguru==0.5.3
uvicorn==0.15.0
python-dotenv==0.19.1
requests==2.25.1